flask-mongo-app/requirements.txt
flask
flask-pymongo
python-dotenv
flasgger